% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimodal.R
\name{dip.test.pval}
\alias{dip.test.pval}
\title{Dip test for univariate unimodalty}
\usage{
dip.test.pval(x, h.sel = kedd::h.mlcv, n.monte.carlo = 1000)
}
\arguments{
\item{x}{univariate data vector}

\item{h.sel}{bandwidth selector. Default to "maximum likelihood cross
validation" selector \code{\link[kedd]{h.mlcv}}.}

\item{n.monte.carlo}{bootstrap sample size}
}
\value{
p-value
}
\description{
Dip test for univariate unimodalty
}
\examples{
\dontrun{
dip.test.pval(rnorm(200))
dip.test.pval(c(rnorm(200), rnorm(200) + 3))}
}
\references{
Hartigan, J. A., & Hartigan, P. M. (1985).
The dip test of unimodality.
\emph{The Annals of Statistics}, 70-84.

M-Y Cheng and Peter Hall.
Calibrating the excess mass and dip tests of modality.
\emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)}, 60(3):579–589, 1998.
}
\seealso{
This function computes a calibrated p-value. This is
different from the (conservative) p-value from \code{\link[diptest]{dip.test}}.
}
