% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{estimate.cdf}
\alias{estimate.cdf}
\title{Estimate the cdf with and without smoothing}
\usage{
estimate.cdf(z)
}
\arguments{
\item{z}{values}
}
\value{
A list consists of functions .ecdf (without smoothing),
.cdf.smoothed (with kernel smoothing) and .x (grid)
}
\description{
Estimate the cdf with and without smoothing
}
\examples{
estimate.cdf(rnorm(30))
}
