% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get.m.out.n.tuples}
\alias{get.m.out.n.tuples}
\title{Generate m-out-of-n subsamples}
\usage{
get.m.out.n.tuples(m, n, B)
}
\arguments{
\item{m}{size of subsample}

\item{n}{size of full sample}

\item{B}{number of subsamples}
}
\value{
A (B x m) matrix where each row is a subsample.
}
\description{
Generate m-out-of-n subsamples
}
\examples{
get.m.out.n.tuples(10, 100, 500)
}
