% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{gof.quantreg.test.single}
\alias{gof.quantreg.test.single}
\title{Goodness-of-fit test of a linear quantile regression model with data splitting}
\usage{
gof.quantreg.test.single(data, tau = 0.5, bsmethod = "xy")
}
\arguments{
\item{data}{data frame, whose column \code{y} should be the outcome}

\item{tau}{quantile}

\item{bsmethod}{bootstrap standard error estimator; see \code{\link[quantreg]{boot.rq}}}
}
\value{
Z-statistic that is N(0,1) distributed when model is well-specified.
When model is misspecified, the statistic tends to be a large positive value.
}
\description{
Test the null hypothesis that \code{y}'s \eqn{\tau}-th quantile equals \eqn{\beta_0 + \beta^{T} X}
for some \eqn{\beta_0} and \eqn{\beta}. That is, this is a specification test
for \code{quantreg::rq(y ~ ., tau=tau, data)}; see \code{\link[quantreg]{rq}}.
}
\note{
This is a randomized, hunt-and-test statistic.
}
\examples{
X <- matrix(rnorm(1000 * 10), ncol=10)
beta <- rep(c(-1,0,1), length.out=10)
# well-specified
y <- 1 + X \%*\% beta + X[,1] * rt(1000, 4)
data <- data.frame(y=y, X=X)
replicate(10, gof.quantreg.test.single(data))
# misspecified
y <- 1 + X \%*\% beta + X[,2]^2 + X[,1] * rt(1000, 4)
data <- data.frame(y=y, X=X)
replicate(10, gof.quantreg.test.single(data))
}
\seealso{
\code{\link[quantreg]{rq}}, \code{\link{test.multisplit}}
}
