% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimodal.R
\name{test.unimodal.dip.hunt.single}
\alias{test.unimodal.dip.hunt.single}
\title{Dip hunting test (single data split) for multivariate linear unimodality}
\usage{
test.unimodal.dip.hunt.single(X, h.sel = kedd::h.mlcv, n.monte.carlo = 1000)
}
\arguments{
\item{X}{Matrix whose rows are iid copies of \eqn{X}}

\item{h.sel}{Bandwidth selector}

\item{n.monte.carlo}{bootstrap sample size}
}
\value{
p-value
}
\description{
Test the null hypothesis that a random vector \eqn{X} comes from a unimodal
distribution, specifically, \emph{linear unimodal} in the sense that
every linear projection of \eqn{X} is unimodal.
}
\details{
The test splits data into two part A and B of equal size. It runs 2-means on
part A to determine a direction. Then it uses part B to conduct \code{\link{dip.test.pval}}.
}
\note{
This is a randomized test.
}
\examples{
\dontrun{
# unimodal
X <- simu.two.balls(n=1000, p=50, sep=0)
replicate(10, test.unimodal.dip.hunt.single(X))
# bimodal
X <- simu.two.balls(n=1000, p=50, sep=2)
replicate(10, test.unimodal.dip.hunt.single(X))}
}
\references{
Hartigan, J. A., & Hartigan, P. M. (1985).
The dip test of unimodality.
\emph{The Annals of Statistics}, 70-84.

M-Y Cheng and Peter Hall.
Calibrating the excess mass and dip tests of modality.
\emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)}, 60(3):579–589, 1998.
}
\seealso{
\code{\link{dip.test.pval}}, \code{\link{test.multisplit}}
}
